Allow to keep the types as they were and just add a default value
It's still possible to omit all unchanged fields in PATCH requests, but
for fields that are given Pydantic will reject data that specifies an
explicit null for fields that aren't optional in the full model.

This is useful when using the package with SQLModel. Without this
feature the client could try to assign a NULL value to a column with a
NOT NULL constraint, which would raise an IntegrityError.